improve hook destructor

+++ src/hooks/use-qr.js
@@ -5,10 +5,14 @@ export function useQr(callback) {
   const [zxingWorker, setZxingWorker] = useState(null);
 
   useEffect(() => {
-    async function createWorkers() {
+    function createWorkers() {
       try {
+        console.info("creating zbar worker");
         const worker1 = new Worker("zbarWorker.js");
+
+        console.info("creating zxing worker");
         const worker2 = new Worker("zxingWorker.js");
+
         const onmessage = (alg) => async ev => {
           if (ev.data != null) {
             const result = ev.data;
@@ -17,8 +21,10 @@ export function useQr(callback) {
             callback({rawcode, milliseconds, alg});
           }
         };
+
         worker1.onmessage = onmessage("zbar");
         worker2.onmessage = onmessage("zxing");
+
         setZbarWorker(worker1);
         setZxingWorker(worker2);
       } catch (err) {
@@ -26,18 +32,15 @@ export function useQr(callback) {
       }
     }
 
-    if (!zbarWorker && !zxingWorker) {
-      createWorkers();
-    }
-    else
-      return function cleanup() {
-        if (zbarWorker && zxingWorker) {
-          console.log("cleaning up worker");
-          zbarWorker.terminate();
-          zxingWorker.terminate();
-        }
-      };
+    createWorkers();
   }, []);
+
+  useEffect(() => {
+    return () => {
+      if (zbarWorker) zbarWorker.terminate();
+      if (zxingWorker) zxingWorker.terminate();
+    }
+  }, [zbarWorker, zxingWorker]);
 
   return [zbarWorker, zxingWorker];
 }
